refactor: 잘못된 계층 위치 수정
